-Format & casts warning fixes -Isolate unused code with ifdefs -Code style corrections -Remove two unneeded NULL casts
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@7249 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/gpre/exp.cpp
@@ -25,7 +25,7 @@
 //
 //____________________________________________________________
 //
-//	$Id: exp.cpp,v 1.12 2003-04-01 11:49:32 brodsom Exp $
+//	$Id: exp.cpp,v 1.13 2003-08-09 18:00:13 brodsom Exp $
 //
 
 #include "firebird.h"
@@ -956,7 +956,7 @@ RSE EXP_rse(GPRE_REQ request, SYM initial_symbol)
 					upcase->nod_arg[0] = item;
 				}
 				count++;
-				PUSH((GPRE_NOD) direction, &directions);
+				PUSH((GPRE_NOD) (ULONG) direction, &directions);
 				if (insensitive)
 					PUSH(upcase, &items);
 				else
